Move functions to planck6_functions file
