+++ keyboards/planck/keymaps/rambohippo/planck6_functions.c
@@ -1,0 +1,188 @@
+#ifndef PLANCK6_FUNCTIONS
+#define PLANCK6_FUNCTIONS
+#include "quantum.h"
+#include "muse.h"
+
+bool is_alt_tab_active = false;
+bool is_ctrl_tab_active = false;
+
+enum planck_layers {
+    _QWERTY,
+    _LOWER,
+    _RAISE,
+    _NUMPAD,
+    _ADJUST
+};
+
+enum planck_keycodes {
+    QWERTY = SAFE_RANGE,
+    LOWER,
+    RAISE,
+    ADJUST,
+    ALT_TAB,
+    CTRLTAB,
+    EXDTVAL
+};
+
+#define NUMPAD MO(_NUMPAD)
+#define ADJUST MO(_ADJUST)
+#define LN_DKFW LCTL(LALT(KC_RGHT))     // Linux - Next Desktop
+#define LN_DKBK LCTL(LALT(KC_LEFT))     // Linux - Previous Desktop
+#define WN_DKFW LCTL(LGUI(KC_RGHT))     // Windows - Next Desktop
+#define WN_DKBK LCTL(LGUI(KC_LEFT))     // Windows - Previous Desktop
+#define WN_DKNW LCTL(LGUI(KC_D))        // Windows - New Desktop
+#define WN_DKCL LCTL(LGUI(KC_F4))       // Windows - Close Desktop
+#define GUI_TAB LGUI(KC_TAB)            // Gui-Tab
+#define CTAB_FW LCTL(KC_TAB)            // Ctrl-Tab
+#define CTAB_BK LSFT(LCTL(KC_TAB))      // Shift-Ctrl-Tab
+#define KC_TERM LGUI(KC_T)              // Super-T - Opens Terminal
+#define BROWSER LGUI(KC_3)              // Gui-3 - Windows - Opens browser
+#define FILES LGUI(KC_E)                // Gui-E - Opens file browser
+
+#ifdef AUDIO_ENABLE
+float plover_song[][2]    = SONG(PLOVER_SOUND);
+float plover_gb_song[][2] = SONG(PLOVER_GOODBYE_SOUND);
+#endif
+
+bool process_record_user(uint16_t keycode, keyrecord_t *record) {
+    switch (keycode) {
+        case QWERTY:
+            if (record->event.pressed) {
+                print("mode just switched to qwerty and this is a huge string\n");
+                set_single_persistent_default_layer(_QWERTY);
+            }
+            return false;
+            break;
+        case LOWER:
+            if (record->event.pressed) {
+                layer_on(_LOWER);
+                update_tri_layer(_LOWER, _RAISE, _ADJUST);
+            } else {
+                layer_off(_LOWER);
+                update_tri_layer(_LOWER, _RAISE, _ADJUST);
+                if (is_alt_tab_active) {
+                    unregister_code(KC_LALT);
+                    is_alt_tab_active = false;
+                }
+                if (is_ctrl_tab_active) {
+                    unregister_code(KC_LCTL);
+                    is_ctrl_tab_active = false;
+                }
+            }
+            return false;
+            break;
+        case RAISE:
+            if (record->event.pressed) {
+                layer_on(_RAISE);
+                update_tri_layer(_LOWER, _RAISE, _ADJUST);
+            } else {
+                layer_off(_RAISE);
+                update_tri_layer(_LOWER, _RAISE, _ADJUST);
+            }
+            return false;
+            break;
+        case ALT_TAB:
+            if (record->event.pressed) {
+                if (!is_alt_tab_active) {
+                    is_alt_tab_active = true;
+                    register_code(KC_LALT);
+                }
+                register_code(KC_TAB);
+            } else {
+                unregister_code(KC_TAB);
+            }
+            break;
+        case CTRLTAB:
+            if (record->event.pressed) {
+                if (!is_ctrl_tab_active) {
+                    is_ctrl_tab_active = true;
+                    register_code(KC_LCTL);
+                }
+                register_code(KC_TAB);
+            } else {
+                unregister_code(KC_TAB);
+            }
+            break;
+        case EXDTVAL:
+            if (record->event.pressed) {
+                tap_code(KC_LALT);
+                tap_code(KC_A);
+                tap_code(KC_V);
+                tap_code(KC_V);
+            }
+            break;
+    }
+    return true;
+}
+
+bool     muse_mode      = false;
+uint8_t  last_muse_note = 0;
+uint16_t muse_counter   = 0;
+uint8_t  muse_offset    = 70;
+uint16_t muse_tempo     = 50;
+
+void encoder_update(bool clockwise) {
+    if (muse_mode) {
+        if (IS_LAYER_ON(_RAISE)) {
+            if (clockwise) {
+                muse_offset++;
+            } else {
+                muse_offset--;
+            }
+        } else {
+            if (clockwise) {
+                muse_tempo += 1;
+            } else {
+                muse_tempo -= 1;
+            }
+        }
+    } else {
+        if (clockwise) {
+#ifdef MOUSEKEY_ENABLE
+            tap_code(KC_MS_WH_DOWN);
+#else
+            tap_code(KC_PGDN);
+#endif
+        } else {
+#ifdef MOUSEKEY_ENABLE
+            tap_code(KC_MS_WH_UP);
+#else
+            tap_code(KC_PGUP);
+#endif
+        }
+    }
+}
+
+void matrix_scan_user(void) {
+#ifdef AUDIO_ENABLE
+    if (muse_mode) {
+        if (muse_counter == 0) {
+            uint8_t muse_note = muse_offset + SCALE[muse_clock_pulse()];
+            if (muse_note != last_muse_note) {
+                stop_note(compute_freq_for_midi_note(last_muse_note));
+                play_note(compute_freq_for_midi_note(muse_note), 0xF);
+                last_muse_note = muse_note;
+            }
+        }
+        muse_counter = (muse_counter + 1) % muse_tempo;
+    } else {
+        if (muse_counter) {
+            stop_all_notes();
+            muse_counter = 0;
+        }
+    }
+#endif
+}
+
+bool music_mask_user(uint16_t keycode) {
+    switch (keycode) {
+        case RAISE:
+        case LOWER:
+            return false;
+        default:
+            return true;
+    }
+}
+
+
+#endif
